ARXIVNG-1422 admin log integration
