+++ core/arxiv/submission/services/classic/tests/test_store_event.py
@@ -1,87 +1,21 @@
-"""
-Integration tests for the classic database service.
-
-These tests assume that SQLAlchemy's MySQL backend is implemented correctly:
-instead of using a live MySQL database, they use an in-memory SQLite database.
-This is mostly fine (they are intended to be more-or-less swappable). The one
-iffy bit is the JSON datatype, which is not available by default in the SQLite
-backend, and so we inject a simple one here. End to end tests with a live MySQL
-database will provide more confidence in this area.
-"""
+"""Tests for storing events."""
 
 from unittest import TestCase, mock
-import os
-from datetime import datetime
-from contextlib import contextmanager
-import json
 
 from flask import Flask
 
-from ...domain.agent import User, System
-from ...domain.submission import License, Submission, Author
-from ...domain.event import CreateSubmission, \
+from ....domain.agent import User, System
+from ....domain.submission import License, Submission, Author
+from ....domain.event import CreateSubmission, \
     FinalizeSubmission, SetPrimaryClassification, AddSecondaryClassification, \
-    SetLicense, SetPrimaryClassification, ConfirmPolicy, \
-    ConfirmContactInformation, SetTitle, SetAbstract, SetDOI, \
-    SetMSCClassification, SetACMClassification, SetJournalReference, \
-    SetComments, SetAuthors, Publish, ConfirmAuthorship, ConfirmPolicy, \
+    SetLicense, ConfirmPolicy, ConfirmContactInformation, SetTitle, \
+    SetAbstract, SetDOI, SetMSCClassification, SetACMClassification, \
+    SetJournalReference, SetComments, SetAuthors, Publish, ConfirmAuthorship, \
     SetUploadPackage
-from . import init_app, create_all, drop_all, models, DBEvent, \
+from .. import init_app, create_all, drop_all, models, DBEvent, \
     get_submission, current_session, get_licenses, exceptions, store_event
 
-
-@contextmanager
-def in_memory_db(app=None):
-    """Provide an in-memory sqlite database for testing purposes."""
-    if app is None:
-        app = Flask('foo')
-    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite://'
-    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-
-    with app.app_context():
-        init_app(app)
-        create_all()
-        try:
-            yield current_session()
-        except Exception:
-            raise
-        finally:
-            drop_all()
-
-
-class TestGetLicenses(TestCase):
-    """Test :func:`.get_licenses`."""
-
-    def test_get_all_active_licenses(self):
-        """Return a :class:`.License` for each active license in the db."""
-        # mock_util.json_factory.return_value = SQLiteJSON
-
-        with in_memory_db() as session:
-            session.add(models.License(
-                name="http://arxiv.org/licenses/assumed-1991-2003",
-                sequence=9,
-                label="Assumed arXiv.org perpetual, non-exclusive license to",
-                active=0
-            ))
-            session.add(models.License(
-                name="http://creativecommons.org/licenses/publicdomain/",
-                sequence=4,
-                label="Creative Commons Public Domain Declaration",
-                active=1
-            ))
-            session.commit()
-            licenses = get_licenses()
-
-        self.assertEqual(len(licenses), 1,
-                         "Only the active license should be returned.")
-        self.assertIsInstance(licenses[0], License,
-                              "Should return License instances.")
-        self.assertEqual(licenses[0].uri,
-                         "http://creativecommons.org/licenses/publicdomain/",
-                         "Should use name column to populate License.uri")
-        self.assertEqual(licenses[0].name,
-                         "Creative Commons Public Domain Declaration",
-                         "Should use label column to populate License.name")
+from .util import in_memory_db
 
 
 class TestStoreEvent(TestCase):
@@ -358,136 +292,3 @@ class TestStoreEvent(TestCase):
         self.assertEqual(db_submission.primary_classification.category,
                          after.primary_classification.category,
                          "Primary classification should be set.")
-
-
-class TestGetSubmission(TestCase):
-    """Test :func:`.get_submission`."""
-
-    def test_get_submission_that_does_not_exist(self):
-        """Test that an exception is raised when submission doesn't exist."""
-        with in_memory_db():
-            with self.assertRaises(exceptions.NoSuchSubmission):
-                get_submission(1)
-
-    def test_get_submission_with_publish(self):
-        """Test that publication state is reflected in submission data."""
-        user = User(12345, 'joe@joe.joe',
-                    endorsements=['physics.soc-ph', 'cs.DL'])
-
-        events = [
-            CreateSubmission(creator=user),
-            SetTitle(creator=user, title='Foo title'),
-            SetAbstract(creator=user, abstract='Indeed' * 10),
-            SetAuthors(creator=user, authors=[
-                Author(order=0, forename='Joe', surname='Bloggs',
-                       email='joe@blo.ggs'),
-                Author(order=1, forename='Jane', surname='Doe',
-                       email='j@doe.com'),
-            ]),
-            SetLicense(creator=user, license_uri='http://foo.org/1.0/',
-                       license_name='Foo zero 1.0'),
-            SetPrimaryClassification(creator=user, category='cs.DL'),
-            ConfirmPolicy(creator=user),
-            SetUploadPackage(creator=user, identifier='12345'),
-            ConfirmContactInformation(creator=user),
-            FinalizeSubmission(creator=user)
-        ]
-
-        with in_memory_db() as session:
-            # User creates and finalizes submission.
-            before = None
-            for i, event in enumerate(list(events)):
-                after = event.apply(before)
-                event, after = store_event(event, before, after)
-                events[i] = event
-                before = after
-            submission = after
-
-            ident = submission.submission_id
-
-            # Moderation happens, things change outside the event model.
-            db_submission = session.query(models.Submission).get(ident)
-
-            # Published!
-            db_submission.status = db_submission.PUBLISHED
-            db_document = models.Document(paper_id='1901.00123')
-            db_submission.document = db_document
-            session.add(db_submission)
-            session.add(db_document)
-            session.commit()
-
-            # Now get the submission.
-            submission_loaded, _ = get_submission(ident)
-
-        self.assertEqual(submission.metadata.title,
-                         submission_loaded.metadata.title,
-                         "Event-derived metadata should be preserved.")
-        self.assertEqual(submission_loaded.arxiv_id, "1901.00123",
-                         "arXiv paper ID should be set")
-        self.assertEqual(submission_loaded.status, Submission.PUBLISHED,
-                         "Submission status should reflect publish action")
-
-    def test_get_submission_with_hold_and_reclass(self):
-        """Test changes made externally are reflected in submission data."""
-        user = User(12345, 'joe@joe.joe',
-                    endorsements=['physics.soc-ph', 'cs.DL'])
-        events = [
-            CreateSubmission(creator=user),
-            SetTitle(creator=user, title='Foo title'),
-            SetAbstract(creator=user, abstract='Indeed' * 20),
-            SetAuthors(creator=user, authors=[
-                Author(order=0, forename='Joe', surname='Bloggs',
-                       email='joe@blo.ggs'),
-                Author(order=1, forename='Jane', surname='Doe',
-                       email='j@doe.com'),
-            ]),
-            SetLicense(creator=user, license_uri='http://foo.org/1.0/',
-                       license_name='Foo zero 1.0'),
-            SetPrimaryClassification(creator=user, category='cs.DL'),
-            ConfirmPolicy(creator=user),
-            SetUploadPackage(creator=user, identifier='12345'),
-            ConfirmContactInformation(creator=user),
-            FinalizeSubmission(creator=user)
-        ]
-
-        with in_memory_db() as session:
-            # User creates and finalizes submission.
-            before = None
-            for i, event in enumerate(list(events)):
-                after = event.apply(before)
-                event, after = store_event(event, before, after)
-                events[i] = event
-                before = after
-            submission = after
-            ident = submission.submission_id
-
-            # Moderation happens, things change outside the event model.
-            db_submission = session.query(models.Submission).get(ident)
-
-            # Reclassification!
-            session.delete(db_submission.primary_classification)
-            session.add(models.SubmissionCategory(
-                submission_id=ident, category='cs.IR', is_primary=1
-            ))
-
-            # On hold!
-            db_submission.status = db_submission.ON_HOLD
-            session.add(db_submission)
-            session.commit()
-
-            # Now get the submission.
-            submission_loaded, _ = get_submission(ident)
-
-        self.assertEqual(submission.metadata.title,
-                         submission_loaded.metadata.title,
-                         "Event-derived metadata should be preserved.")
-        self.assertEqual(submission_loaded.primary_classification.category,
-                         "cs.IR",
-                         "Primary classification should reflect the"
-                         " reclassification that occurred outside the purview"
-                         " of the event model.")
-        self.assertEqual(submission_loaded.status, Submission.ON_HOLD,
-                         "Submission status should still be submitted.")
-        self.assertTrue(submission_loaded.is_on_hold,
-                        "Hold status should reflect hold action performed"
-                        " outside the purview of the event model.")
